refactor(wallet_policy): move WalletPolicy into a private policy module
- the fields and validate() are now private to the submodule, so every
  construction path validates
- the TryFrom<&str> template branch goes through a private from_template
  instead of a struct literal
